debug and dispersal
Created a "2-quality" version of the new brunswick field site. Use temp_dispersal to work with it.

small_tester.R is for getting new features up and running.

working on implementing dispersal. In this commit I clarified that compute_sparse_distance is only for distances within a max distance. No more average dispersal as an argument. This matrix is then used to calculate the dispersal kernel, initialized in compute_dispersal. What's here is not working and is from LLM as a scaffold.